+++ src/Cashout/Transfer.php
@@ -108,7 +108,7 @@ class Transfer extends AbstractOperationProcessor
                     array('miraklId' => $operation->getMiraklId(), "action" => "Transfer")
                 );
             } catch (Exception $e) {
-                $this->logger->warning(
+                $this->logger->critical(
                     "[KO] Transfer operation failed",
                     array('miraklId' => $operation->getMiraklId(), "action" => "Transfer")
                 );
@@ -131,7 +131,11 @@ class Transfer extends AbstractOperationProcessor
         try {
             $vendor = $this->getVendor($operation);
 
-            if (!$vendor || $this->hipay->isAvailable($vendor->getEmail())) {
+            if (!$vendor) {
+                throw new WalletNotFoundException($vendor);
+            }
+
+            if (!$this->hipay->isWalletExist($vendor->getHiPayId())) {
                 throw new WalletNotFoundException($vendor);
             }
 
@@ -156,7 +160,7 @@ class Transfer extends AbstractOperationProcessor
             //Transfer
             $transferId = $this->hipay->transfer($transfer, $vendor);
 
-            $operation->setStatus(new Status(Status::TRANSFER_SUCCESS));
+            $operation->setStatus(new Status(Status::TRANSFER_REQUESTED));
             $operation->setTransferId($transferId);
             $operation->setUpdatedAt(new DateTime());
             $this->operationManager->save($operation);
@@ -164,7 +168,7 @@ class Transfer extends AbstractOperationProcessor
             $this->logOperation(
                 $operation->getMiraklId(),
                 $operation->getPaymentVoucher(),
-                Status::TRANSFER_SUCCESS,
+                Status::TRANSFER_REQUESTED,
                 ""
             );
 
@@ -241,5 +245,4 @@ class Transfer extends AbstractOperationProcessor
 
         return $toTransfer;
     }
-
 }