+++ tests/Notification/HandlerTest.php
@@ -1,12 +1,15 @@
 <?php
+
 namespace HiPay\Wallet\Mirakl\Test\Notification;
+
 use DateTime;
 use HiPay\Wallet\Mirakl\Cashout\Model\Operation\Status;
-use HiPay\Wallet\Mirakl\Test\Notification\HandlerNotif;
+use HiPay\Wallet\Mirakl\Notification\Handler;
 use HiPay\Wallet\Mirakl\Test\Common\AbstractProcessorTest;
 use HiPay\Wallet\Mirakl\Test\Stub\Entity\Operation;
 use HiPay\Wallet\Mirakl\Test\Stub\Entity\Vendor;
 use Prophecy\Argument;
+
 /**
  *
  * @author    Ivanis Kouamé <ivanis.kouame@smile.fr>
@@ -18,29 +21,91 @@ class HandlerTest extends AbstractProcessorTest
     protected $notificationHandler;
     protected $testFilesPath;
     protected $notificationEventClassPath;
+
     public function setUp()
     {
         parent::setUp();
+
         $this->testFilesPath = __DIR__ . "/../../data/test/notification/";
-        $this->notificationHandler = new HandlerNotif(
+
+        $this->notificationHandler = new Handler(
             $this->eventDispatcher->reveal(),
             $this->logger->reveal(),
             $this->operationManager->reveal(),
-            $this->vendorManager->reveal()
+            $this->vendorManager->reveal(),
+            $this->logVendorManager->reveal(),
+            $this->apiFactory->reveal(),
+            $this->logOperationsManager->reveal()
         );
     }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
      * @throws \HiPay\Wallet\Mirakl\Exception\IllegalNotificationOperationException
+     * @expectedException \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
      */
     public function testMd5Failure()
     {
         $xml = $this->readFile("md5Fail.xml");
-        $this->setExpectedException("\\HiPay\\Wallet\\Mirakl\\Exception\\ChecksumFailedException");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
         $this->eventDispatcher->dispatch(Argument::any())->shouldNotBeCalled();
+
         $this->notificationHandler->handleHiPayNotification($xml);
     }
+
+    /**
+     * @cover ::handleHiPayNotification
+     */
+    public function testMd5Different()
+    {
+        $xml = $this->readFile("other.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "testfalse",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $this->vendorManager->save(Argument::any())->shouldBeCalled();
+
+        $this->notificationHandler->handleHiPayNotification($xml);
+    }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -49,9 +114,32 @@ class HandlerTest extends AbstractProcessorTest
     public function testBankInfoNotification()
     {
         $xml = $this->readFile("bankInfoValidation.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
         $this->setEventAssertion("bankInfos", "BankInfo");
+
         $this->notificationHandler->handleHiPayNotification($xml);
     }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -60,9 +148,31 @@ class HandlerTest extends AbstractProcessorTest
     public function testOtherNotification()
     {
         $xml = $this->readFile("other.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
         $this->setEventAssertion("other", "Other");
         $this->notificationHandler->handleHiPayNotification($xml);
     }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -71,15 +181,24 @@ class HandlerTest extends AbstractProcessorTest
     public function testIdentificationNotification()
     {
         $xml = $this->readFile("identification.xml");
-        $vendor = new Vendor('test@test', 123456, null, null, false);
-        $this->vendorManager->findByHiPayId(123456)->willReturn($vendor)->shouldBeCalledTimes(1);
-        $this->vendorManager->save(Argument::that(function($vendorUpdated) {
+
+        $vendor = new Vendor('test@test', 123456, null, null, false, 1, "test");
+
+        $this->vendorManager->findByHiPayId(123456)->willReturn($vendor)->shouldBeCalledTimes(2);
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $this->vendorManager->save(Argument::that(function ($vendorUpdated) {
             // Should have been updated to identified
             return $vendorUpdated->getHipayIdentified() === true;
         }))->shouldBeCalledTimes(1);
+
         $this->setEventAssertion("identification", "Identification");
         $this->notificationHandler->handleHiPayNotification($xml);
     }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -88,14 +207,217 @@ class HandlerTest extends AbstractProcessorTest
     public function testWithdrawSuccessNotification()
     {
         $xml = $this->readFile("withdrawSuccess.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
         $operation = new Operation(2000, new DateTime(), "000001", rand());
+
         $operation->setStatus(new Status(Status::WITHDRAW_REQUESTED));
-        $this->operationManager->findByWithdrawalId(Argument::type("string"))->willReturn($operation)->shouldBeCalled();
+
+        $this->operationManager
+            ->findByWithdrawalId(Argument::type("string"))
+            ->willReturn($operation)
+            ->shouldBeCalled();
+
         $this->operationManager->save(Argument::is($operation))->shouldBeCalled();
+
         $this->setEventAssertion(array("withdraw", "success"), "Withdraw");
+
         $this->notificationHandler->handleHiPayNotification($xml);
+
         $this->assertEquals(Status::WITHDRAW_SUCCESS, $operation->getStatus());
     }
+
+    /**
+     * @cover ::handleHiPayNotification
+     */
+    public function testTransferSuccessAuthorize()
+    {
+        $xml = $this->readFile("authorizeSuccess.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $operation = new Operation(2000, new DateTime(), "000001", rand());
+
+        $operation->setStatus(new Status(Status::TRANSFER_REQUESTED));
+
+        $this->operationManager
+            ->findOneByTransferId(Argument::type("string"))
+            ->willReturn($operation)
+            ->shouldBeCalled();
+
+        $this->operationManager->save(Argument::is($operation))->shouldNotBeCalled();
+
+        $this->notificationHandler->handleHiPayNotification($xml);
+
+        $this->assertEquals(Status::TRANSFER_REQUESTED, $operation->getStatus());
+    }
+
+    /**
+     * @cover ::handleHiPayNotification
+     */
+    public function testTransferFailureAuthorize()
+    {
+        $xml = $this->readFile("authorizeFailure.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $operation = new Operation(2000, new DateTime(), "000001", rand());
+
+        $operation->setStatus(new Status(Status::TRANSFER_REQUESTED));
+
+        $this->operationManager
+            ->findOneByTransferId(Argument::type("string"))
+            ->willReturn($operation)
+            ->shouldBeCalled();
+
+        $this->operationManager->save(Argument::is($operation))->shouldBeCalled();
+
+        $this->notificationHandler->handleHiPayNotification($xml);
+
+        $this->assertEquals(Status::TRANSFER_FAILED, $operation->getStatus());
+    }
+
+    /**
+     * @cover ::handleHiPayNotification
+     */
+    public function testTransferCaptureSuccess()
+    {
+        $xml = $this->readFile("captureSuccess.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $operation = new Operation(2000, new DateTime(), "000001", rand());
+
+        $operation->setStatus(new Status(Status::TRANSFER_REQUESTED));
+
+        $this->operationManager
+            ->findOneByTransferId(Argument::type("string"))
+            ->willReturn($operation)
+            ->shouldBeCalled();
+
+        $this->operationManager->save(Argument::is($operation))->shouldBeCalled();
+
+        $this->notificationHandler->handleHiPayNotification($xml);
+
+        $this->assertEquals(Status::TRANSFER_SUCCESS, $operation->getStatus());
+    }
+
+    /**
+     * @cover ::handleHiPayNotification
+     */
+    public function testTransferCaptureFailure()
+    {
+        $xml = $this->readFile("captureFailure.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $operation = new Operation(2000, new DateTime(), "000001", rand());
+
+        $operation->setStatus(new Status(Status::TRANSFER_REQUESTED));
+
+        $this->operationManager
+            ->findOneByTransferId(Argument::type("string"))
+            ->willReturn($operation)
+            ->shouldBeCalled();
+
+        $this->operationManager->save(Argument::is($operation))->shouldBeCalled();
+
+        $this->notificationHandler->handleHiPayNotification($xml);
+
+        $this->assertEquals(Status::TRANSFER_FAILED, $operation->getStatus());
+    }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -104,14 +426,45 @@ class HandlerTest extends AbstractProcessorTest
     public function testWithdrawCancelNotification()
     {
         $xml = $this->readFile("withdrawCanceled.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
         $operation = new Operation(2000, new DateTime(), "000001", rand());
+
         $operation->setStatus(new Status(Status::WITHDRAW_REQUESTED));
-        $this->operationManager->findByWithdrawalId(Argument::type("string"))->willReturn($operation)->shouldBeCalled();
+
+        $this->operationManager
+            ->findByWithdrawalId(Argument::type("string"))
+            ->willReturn($operation)
+            ->shouldBeCalled();
+
         $this->operationManager->save(Argument::is($operation))->shouldBeCalled();
+
         $this->setEventAssertion(array("withdraw", "canceled"), "Withdraw");
+
         $this->notificationHandler->handleHiPayNotification($xml);
+
         $this->assertEquals(Status::WITHDRAW_CANCELED, $operation->getStatus());
     }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -120,9 +473,32 @@ class HandlerTest extends AbstractProcessorTest
     public function testUnknownOperation()
     {
         $xml = $this->readFile("unknownOperation.xml");
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
+        );
+
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
         $this->eventDispatcher->dispatch(Argument::any())->shouldNotBeCalled();
+
         $this->notificationHandler->handleHiPayNotification($xml);
     }
+
     /**
      * @cover ::handleHiPayNotification
      * @throws \HiPay\Wallet\Mirakl\Exception\ChecksumFailedException
@@ -131,20 +507,32 @@ class HandlerTest extends AbstractProcessorTest
     public function testDocumentValidation()
     {
         $xml = $this->readFile("documentValidation.xml");
-        $this->eventDispatcher->dispatch(Argument::any())->shouldNotBeCalled();
-        $parameters = array(
-            'mail.host' => 'smtp',
-            'mail.port' => '1025',
-            'mail.security' => null,
-            'mail.username' => null,
-            'mail.password' => null,
-            'mail.subject'=> 'Mirakl HiPay Connector Notification',
-            'mail.to' => 'marketplace.operator@hipay.com',
-            'mail.from' => 'mirakl.hipay.connector@hipay.com',
+
+        $vendor = new Vendor(
+            "test@test.com",
+            1,
+            1,
+            1,
+            1,
+            1,
+            "test",
+            1,
+            'FR'
         );
 
-        $this->notificationHandler->handleHiPayNotification($xml, $parameters);
+        $this->vendorManager->findByHiPayId(Argument::any())
+            ->willReturn($vendor)
+            ->shouldBeCalled();
+
+        $this->hipay->getAccountHiPay(Argument::any())
+            ->willReturn(array("callback_salt" => "test"))
+            ->shouldBeCalled();
+
+        $this->eventDispatcher->dispatch(Argument::any())->shouldNotBeCalled();
+
+        $this->notificationHandler->handleHiPayNotification($xml);
     }
+
     /**
      * Read a test file
      *
@@ -157,6 +545,7 @@ class HandlerTest extends AbstractProcessorTest
         $xml = file_get_contents($path);
         return $xml;
     }
+
     /**
      * Add an assertion about the event dispatcher
      *
@@ -165,7 +554,8 @@ class HandlerTest extends AbstractProcessorTest
      */
     protected function setEventAssertion($eventNameParts, $eventClass)
     {
-        $eventNameParts = (array) $eventNameParts;
+        $eventNameParts = (array)$eventNameParts;
+
         $eventNameArgument = Argument::that(function ($argument) use ($eventNameParts) {
             $part = reset($eventNameParts);
             do {
@@ -175,12 +565,9 @@ class HandlerTest extends AbstractProcessorTest
             } while (false !== next($eventNameParts));
             return true;
         });
+
         $notificationEventClassPath = "\\HiPay\\Wallet\\Mirakl\\Notification\\Event\\";
         $eventTypeArgument = Argument::type($notificationEventClassPath . $eventClass);
-        $this->eventDispatcher->dispatch(
-            $eventNameArgument,
-            $eventTypeArgument
-        )
-            ->shouldBeCalled();
+        $this->eventDispatcher->dispatch($eventNameArgument, $eventTypeArgument)->shouldBeCalled();
     }
 }